+++ ui/tool_liquify.py
@@ -46,7 +46,7 @@ class tool_liquify:
 	
 	def liquify(self, origin):
 		#sacrifice anti-alias for speed
-		return origin[self.mask_y.astype(np.uint8),self.mask_x.astype(np.uint8)]
+		return origin[self.mask_y.astype(np.int),self.mask_x.astype(np.int)]
 
 	def update(self, image, mask, points):
 		num_pnts = len(points)
